+++ src/main/java/fr/proneus/engine/input/MouseScrollManager.java
@@ -6,31 +6,31 @@ import fr.proneus.engine.Game;
 import fr.proneus.engine.gui.Component;
 import fr.proneus.engine.state.State;
 
-public class MouseScrollManager extends GLFWScrollCallback{
+public class MouseScrollManager extends GLFWScrollCallback {
 
-	private Game game;
-	
-	private State state;
-	
-	public MouseScrollManager(Game game) {
-		this.game = game;
-	}
-	
-	@Override
-	public void invoke(long window, double x, double y) {
-		
-		
-		state.onMouseScroll(game, y);
-		for (Component comp : state.getComponents()) {
-			if (comp.isVisible())
-				comp.onMouseScroll(game, y);
-		}
-		
-	}
-	
-	public void setListener(State state) {
-		this.state = state;
-		
-	}
+    private Game game;
+
+    private State state;
+
+    public MouseScrollManager(Game game) {
+        this.game = game;
+    }
+
+    @Override
+    public void invoke(long window, double x, double y) {
+
+
+        state.onMouseScroll(game, (float) y);
+        for (Component comp : state.getComponents()) {
+            if (comp.isVisible())
+                comp.onMouseScroll(game, y);
+        }
+
+    }
+
+    public void setListener(State state) {
+        this.state = state;
+
+    }
 
 }
